+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ## Release 2022-11-11
 - Fixed a bug in ElectionDay calendar, was incorrectly calculating a date on 1-Nov 
 which is not allowed
-- Set up the Ercot ISO.  Added Ercot specific buckets.
+- Set up the Ercot ISO.  Added Ercot specific buckets.  
 
 ## Release 2022-10-31
 - Add a name to the holiday type enum